refactor: store analyticsLink not isSubmissionService
